+++ packages/codegen-ui-react/lib/forms/form-renderer-helper/cta-props.ts
@@ -34,6 +34,7 @@ import { isManyToManyRelationship } from './map-from-fieldConfigs';
 
 export const buildDataStoreExpression = (
   dataStoreActionType: 'update' | 'create',
+  modelName: string,
   importedModelName: string,
   fieldConfigs: Record<string, FieldConfigMetadata>,
 ) => {
@@ -148,7 +149,7 @@ export const buildDataStoreExpression = (
                     ),
                     undefined,
                     [
-                      factory.createIdentifier(`${lowerCaseFirst(importedModelName)}Record`),
+                      factory.createIdentifier(`${lowerCaseFirst(modelName)}Record`),
                       factory.createArrowFunction(
                         undefined,
                         undefined,
@@ -202,29 +203,6 @@ export const buildDataStoreExpression = (
         ),
       ]
     : [
-        factory.createVariableStatement(
-          undefined,
-          factory.createVariableDeclarationList(
-            [
-              factory.createVariableDeclaration(
-                factory.createIdentifier('original'),
-                undefined,
-                undefined,
-                factory.createAwaitExpression(
-                  factory.createCallExpression(
-                    factory.createPropertyAccessExpression(
-                      factory.createIdentifier('DataStore'),
-                      factory.createIdentifier('query'),
-                    ),
-                    undefined,
-                    [factory.createIdentifier(importedModelName), factory.createIdentifier('id')],
-                  ),
-                ),
-              ),
-            ],
-            NodeFlags.Const,
-          ),
-        ),
         factory.createExpressionStatement(
           factory.createAwaitExpression(
             factory.createCallExpression(
@@ -241,7 +219,7 @@ export const buildDataStoreExpression = (
                   ),
                   undefined,
                   [
-                    factory.createIdentifier('original'),
+                    factory.createIdentifier(`${lowerCaseFirst(modelName)}Record`),
                     factory.createArrowFunction(
                       undefined,
                       undefined,